packages/middlewares/image-processors/hashing-image-processor/src/lambdas/processor/requirements.txt
boto3
aws_lambda_powertools
aws-xray-sdk
imagehash==4.3.1
pillow==10.3.0
